Weak linking for Arduino (#10)
Adds a weak linkage for the gpio_init function, which Arduino Mbed OS hides away as _gpio_init because of "conflicting names", despite the fact that there wouldn't be a conflict because of the different call signatures. Either way, this weak linkage is discarded in case of using Earle Philhower's core, or when using the Pico SDK and only sticks when using the standard Arduino core (Mbed OS).